Auf strengere Warnings angepasst.
git-svn-id: file:///e/Backups/vorderklier.de/backup/160817/data/svn/PartyDJ_Alt@319 0e6492ef-cedd-114b-947f-3769be7ed64a

+++ src/common/Functions.java
@@ -15,7 +15,7 @@ public class Functions
 	 */
 	public static String formatTime(double time)
 	{
-        long seconds = (long)Math.round(time);
+        long seconds = Math.round(time);
         long hours = seconds / 3600;
         seconds = seconds % 3600;
         long minutes = seconds / 60;
